Crate libcommon

Re-exports

pub use paste;

Modules

bindings

c_main_waker

Waker infrastructure to drive a Future task to completion on the main C thread.

farch

Unarchive embedded content in C files.

helpers

Utility helpers functions.

iop

Module to interact with IOPs in Rust

lstr

Type-safe wrappers for C lstr_t with ownership and content type safety.

mem_stack

t_pool implementation and manipulation in Rust.

pstream

Conversion between a C pstream_t and Rust string/bytes types.

run_callback

Infrastructure to run functions that need callback in main C event loop

sb

Module to interact with sb_t in Rust.

thr

Module to export the thread functions used by libcommon in a safe way

Macros

iop_get

Get a field value from an IOP struct or union.

iop_new

Create a new IOP struct/class with optional field initialization, or a union with a required variant.

iop_set

Set multiple fields on an IOP struct or union.