libcommon::bindings

Function __iop_jpack_file

pub unsafe extern "C" fn __iop_jpack_file(
    iop_env_ctx: *const iop_env_ctx_t,
    filename: *const c_char,
    file_flags: c_uint,
    file_mode: mode_t,
    st: *const iop_struct_t,
    value: *const c_void,
    flags: c_uint,
    subfiles: *const qv_iop_json_subfile_t,
    err: *mut sb_t,
) -> c_int

Serialize an IOP C structure in an IOP-JSon file.

This function packs an IOP structure into (strict) JSon format and writes it in a file.

Some IOP sub-objects can be written in separate files using the include feature. Only one level of inclusion is supported.

\param[in] iop_env_ctx The IOP environment. \param[in] filename The file in which the value is packed. \param[in] file_flags The flags to use when opening the file (\ref enum file_flags). \param[in] file_mode The mode to use when opening the file. \param[in] st IOP structure description. \param[in] value Pointer on the IOP structure to pack. \param[in] flags Packer flags bitfield (see iop_jpack_flags). \param[in] subfiles If set, this is the list of IOP objects that must be written in separate files using @include. \param[out] err Buffer filled in case of error. \param[out] err NULL or the buffer to use to write textual error.