libcommon::bindings

Function el_timer_register_d

pub unsafe extern "C" fn el_timer_register_d(
    next: i64,
    repeat: i64,
    flags: ev_timer_flags_t,
    arg1: Option<unsafe extern "C" fn(*mut ev_t, data_t)>,
    arg2: data_t,
) -> *mut ev_t

\brief registers a timer

There are two kinds of timers: one shot and repeating timers:

One shot timers fire their callback once, when they time out.

Repeating timers automatically rearm after being fired.

One shot timers are automatically destroyed at the end of the callback if they have not be rearmed in it. As a consequence, you must be careful to cleanup all references to one-shot timers you don’t rearm within the callback.

\param[in] next relative time in ms at which the timers fires. \param[in] repeat repeat interval in ms, 0 means single shot. \param[in] flags timer related flags (nomiss, lowres, …) \param[in] cb callback to call upon timer expiry. \param[in] priv private data. \return the timer handler descriptor.