libcommon::bindings

Function iop_enum_from_str_desc

pub unsafe extern "C" fn iop_enum_from_str_desc(
    ed: *const iop_enum_t,
    s: *const c_char,
    len: c_int,
    err: c_int,
) -> c_int

Convert a string to its integer value using an IOP enum mapping.

This function will return err if the string value doesn’t exist in the enum set.

Prefer the macro version iop_enum_from_str() instead of this low-level API.

\param[in] ed The IOP enum definition (__e). \param[in] s String value to look for. \param[in] len String length (or -1 if unknown). \param[in] err Value to return in case of conversion error.