libcommon::bindings

Function mpsc_queue_drain_start

pub unsafe extern "C" fn mpsc_queue_drain_start(
    it: *mut mpsc_it_t,
    q: *mut mpsc_queue_t,
)

\brief initiates the iterator to start a drain.

\warning it’s forbidden not to drain fully after this function has been called.

a typical drain looks like that:

  static void doit(mpsc_node_t *node, data_t data)
  {
      process_node(node);
      freenode(node);
  }

  {
      mpsc_it_t it;

      mpsc_queue_drain_start(&it, &path->to.your_queue);
      do {
          mpsc_node_t *n = mpsc_queue_drain_fast(&it, doit,
                                                 (data_t){ .ptr = NULL });
          process_node(n);
          // XXX do NOT freenode(n) here mpsc_queue_drain_end will use it
      } while (!mpsc_queue_drain_end(&it, freenode));
  }

It is important to note that the queue may go through the draining loop multiple times if jobs are enqueued while the queue is drained.

\param[out] it the iterator to initialize \param[in] q the queue to drain