libcommon::bindings

Function t_yaml_pack

pub unsafe extern "C" fn t_yaml_pack(
    env: *mut yaml_pack_env_t,
    data: *const yaml_data_t,
    writecb: yaml_pack_writecb_f,
    priv_: *mut c_void,
    err: *mut sb_t,
) -> c_int

Pack a YAML data.

The callback \p writecb will be called for every buffer than must be written.

Unless you need to write a custom writer, you probably want to use yaml_pack_sb or yaml_pack_file instead.

\param[in] env Packing environment. \param[in] data The YAML data to pack. \param[in] presentation Presentation data to reformat the document properly. If NULL, the presentation data that may be present in \p data is used. \param[in] writecb Callback called on every buffer that must be written. \param[in] data Private data passed to \p writecb. \param[out] err Buffer filled in case of error.