libcommon::helpers

Function slice_assume_init_mut

pub const unsafe fn slice_assume_init_mut<T>(
    slice: &mut [MaybeUninit<T>],
) -> &mut [T]

Gets a mutable (unique) reference to a maybe uninitialized slice.

Warning

This function assumes that the slice is already initialized. Using this function on an uninitialized slice is undefined behavior and can lead to safety issues, crashes, or other unpredictable behaviors. Do not use it to initialize the slice; use a mutable pointer instead.

Safety

The caller must ensure that all elements in the slice are fully initialized before calling this function. Calling this function with an uninitialized slice is undefined behavior.

TODO: use slice::assume_init_mut once stabilized.