libcommon::helpers

Function slice_from_nullable_raw_parts

pub const unsafe fn slice_from_nullable_raw_parts<'a, T>(
    data: *const T,
    len: usize,
) -> &'a [T]

Create a slice from a potential null pointer and length.

std::slice::from_raw_parts() must be used with a non-null pointer. If the pointer is null, use a dandling non-null pointer.

Safety

See from_raw_parts.html#safety, except that data can be null.

Panics

len must be 0 if data is null.