libcommon::lstr

Trait AsRawLstr

pub trait AsRawLstr {
    // Required method
    fn as_raw_lstr(&self) -> lstr_t;
}

Trait to convert a value to a raw lstr_t.

This trait is implemented by all lstr wrapper types and allows generic code to work with any lstr type when calling C functions.

Required Methods

fn as_raw_lstr(&self) -> lstr_t

Create a raw lstr_t from this value.

Note: The lifetime of the underlying data is lost in the returned lstr_t. The caller must ensure the returned lstr_t does not outlive the source. Creating the raw lstr_t is safe (like creating a raw pointer), but using it to access data requires appropriate unsafe blocks.